Move 'chart' to components, since it seems a pretty bare, building-block kind of component.
